requirements.txt
streamlit
Pillow
reportlab
requests
deep-translator
